pyproject.toml
[project]
name = "dpr"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.7.1",
    "aiohttp>=3.12.13",
    "audio-recorder-streamlit>=0.0.10",
    "fastapi>=0.115.14",
    "fastmcp>=2.10.1",
    "groq>=0.29.0",
    "mcp>=1.10.1",
    "numba>=0.61",
    "numpy<2.3",
    "openai>=1.93.0",
    "openai-agents[viz,voice]>=0.1.0",
    "openai-whisper>=20250625",
    "openpyxl>=3.1.5",
    "pandas>=2.3.0",
    "psutil>=7.0.0",
    "pyaudio>=0.2.14",
    "pydantic>=2.11.7",
    "pydantic-ai>=0.3.6",
    "pydantic-ai-slim[groq]>=0.3.6",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "sounddevice>=0.5.2",
    "speechrecognition>=3.14.3",
    "streamlit-extras>=0.7.5",
    "watchdog>=6.0.0",
]
